Restructure dir;nest new-class dir in my-classes
